Split a function creating a container into few parts.
This is the main function for creating the container. Call sites are
converted in env.c. Some of the work will actually be reused by upstream
linux containers, such as kicking /sbin/init, and some of the basic
setup. Those functions are moved to separate functions that will be
called later on from upstream linux creation routine.

[kir@openvz.org: patch splitting]
Signed-off-by: Glauber Costa <glommer@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>