move set_personality32 calls to common code.
Since this is also needed in upstream Linux kernels container creation,
we move the corresponding call to one of the common functions executed
during env create. Not all calls can be done in common code, though.
OpenVZ still sets personality after every issue of the create ioctl,
which is itself done in various parts of the code. The definition can
still be in a common location.

[kir@: trivial nitpicks]

Signed-off-by: Glauber Costa <glommer@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>