basic libcgroup support
This patch lays out the basic cgroup support for vzctl. Configuration
is not yet applied. Some cgroup have mandatory that needs to be done
in order for it to run (cpusets, for instance, need particular masks
to be set, or it won't function). That is done manually, but in the
future, should be replaced with a merger between configuration file data
and mandatory data.

[kir@: configure.ac rework, minor nitpicks]

Signed-off-by: Glauber Costa <glommer@parallels.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>