is_ploop_supported(): reimplement using /proc/vz/ploop_minor
Since ploop library drastically changed the ploop_getminor() function
(and made it internal) we can't use it to check if ploop is supported
by the running kernel.

Instead, let's check the existense of /proc/vz/ploop_minor, which is
available since kernel 042stab058.7. If it is not available, either
kernel do not support ploop, or it's the old kernel (read please upgrade).

Fix ploop-lib version requirements accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>