Implement upstream veth function
We use veth functions to add a virtual ethernet device inside
container (veth).

This is a simple mode of operation that creates a pair of virtual
ethernet devices, one in the host, the other in the container.

No further setup is done besides having the devices created with the
relevant options.

We'll use the available ip command to do a huge part of our network
configuration. This patch adds the necessary scripts and changes to
existing scripts to do it.

[kir@: minor nitpicks]

Signed-off-by: Glauber Costa <glommer@parallels.com>