Implement upstream netdev function
We use netdev_ctl function to deal with a network device inside
container. Those would be commands like --netdev-add and its respective
deletion.

The goal of netdev is to provide the container with a device that
already exists in the box. We'll use the ip utility to achieve that, by
moving the said interface to the network namespace the container runs
at.

Signed-off-by: Glauber Costa <glommer@parallels.com>