Allow to build w/o ploop headers
Use HAVE_PLOOP defines for automake and C to eliminate ploop-related code.

Note we still call is_ploop_supported(), which, in case of no HAVE_PLOOP defined,
tells (only once) that ploop support is not compiled in.

NOTE if you want to recompile after changing --with-ploop to --without-ploop or
vice versa, you need to do make clean first. This is because C defines are coming
in the command line (i.e. we don't have acconfig.h) so make can't detect that
dependencies have changed and rebuild the appropriate stuff.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>