vzctl set --mount_opts: warn if changing on running CT
In fact, we could try doing remounting with new options, but it's
a bit problematic since some mount options are set inside ploop lib so
we have to get the old mount options, merge those with the new ones and
then do remount. I am leaving it as a homework for someone who needs
changing mount options runtime.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>