vzctl.spec: split the package into vzctl-core and vzctl
Since vzctl is now capable of running on top of non-openvz kernel,
having OpenVZ-specific scripts and requirements (dependencies on vzkernel,
ploop etc) in vzctl package is questionable.

The solution is to split the beast into vzctl-core (with as little
requirements as we can) and vzctl (with all the rest).

From now on, installing vzctl-core will give you an ability to run
upstream linux containers, and installing vzctl will give you the rest.

Eventually some tools (such as vzlist) will be ported to support
upstream kernel -- and then they will be moved on to vzctl-core.
Same about initscript, vzeventd, and what-else-do-we-have-here.

Note that vzctl-lib is gone as a separate package, it is now included
into vzctl-core.

This patch also enables building vzctl for upstream (adding --with-cgroup
to ./configure).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>