vzctl start/stop: print error for non-applicable options
parse_startstop_opt() is a bit tricky because it tries to parse both start
and stop options. Existing code properly rejects a bad option and returns
the proper error code, but it does that silently. That leads to e.g:

	# vzctl start 101 --skip-umount --force
	# echo $?
	20

Fix it by providing an error message. Now it is:

	# vzctl start 101 --skip-umount --force
	Option --skip-umount is not applicable to start command

Signed-off-by: Kir Kolyshkin <kir@openvz.org>