configure.ac: check for $target_cpu not $build_cpu
What we should care for is $target_cpu. It is not important in
the usual not cross-compiling case, but let's be perfect.

While at it, (re-)add a comment explaining why we need -m64. This
comment was initially introduced by commit 1eb3cfa and was later lost
in commit d69917c.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>