configure.ac: for x86_64, set libdir to lib64
If building for x86_64 and libdir is not explicitly set, set it to
end up in lib64 as it should be. This should result in usable vzctl
when doing manual compile/install.

Note that when we build rpms, we explicitly set --libdir when calling
configure, but if we try a manual build we might end up installing
libraries into wrong place.

Reported-by: Glauber Costa <glommer@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>