Change mount to a symlink for /var/run/netns file
Although the documentation for ip suggests that the file
in /var/run/netns should be a bind mount of the /proc file, it also
claims that:

1) Any fd will do
2) This will keep the reference to the namespace open until one umounts.

We don't really need to assert 2). After the container is dead, we have
no business with its namespace. If the container dies, for instance,
without having the chance to clean itself up, it is actually better that
we don't hold an extra reference. So change it to a symlink.

Thanks to Andrew Vagin for detecting the problem, and suggesting the
solution.

Acked-by: Andrew Vagin <avagin@parallels.com>
Signed-off-by: Glauber Costa <glommer@parallels.com>