Do not enclose all containers in a vz directory
It could be interesting to keep the vz/ directory in cgroups
filesystem; both for organization and to allow admins to apply
restrictions to the whole set of containers. That is why we have been
doing it this way so far.

There is, however, general upstream acceptance that this should be
avoided. Although it seems to the outside as a merely grouping facility,
there are costs associated with this that can get high, besides not
always being easy to grasp.

Given the trade-offs, it is better for us to create directories in the
form vz-xxx, with all containers in the same level.

Signed-off-by: Glauber Costa <glommer@parallels.com>