change header of cgroup configuration function
We were deadling with mostly long quantities up to now. When we start
dealing with the device cgroup, it is a lot more convenient to pass a
string (passing the struct dev directly would also work).

So as a first step, change the function signature to accept a void
pointer instead. We provide a val = _val void-to-ulong translation to
minimize disruption to the rest of the code

Signed-off-by: Glauber Costa <glommer@parallels.com>