vzctl: hide '/usr/sbin/vzquota not found' message
vzquota can only be used for OpenVZ kernel and simfs layout.
In most places we fixed it by disabling quota for upstream
case, but here in fsumount we should try to turn off
quota regardless of the setting (to be able to turn off
quota on a container for which we disabled it after start).

Now, with vzctl-core vzctl can be used on upstream containers
without vzquota, and vzquota binary is not available, so we
receive this bogus error:

	# vzctl create 111
	....
	Creating container private area (centos-5-x86-devel)
	Performing postcreate actions
>>>>	File /usr/sbin/vzquota not found
	CT configuration saved to /etc/vz/conf/111.conf
	Container private area was created

Fix it by checking for vzquota binary existence

Signed-off-by: Kir Kolyshkin <kir@openvz.org>