cleanup container before startup
Right now, we can leak some values from one run to another, if for any
reason this container is still represented in the cgroup filesystem.
This means that one-time configuration from the previous run can
potentially stay, specially if there aren't any saved value for the
said property.

A robust and conservative approach is to make sure that everything
related to our container is always cleaned up before we start a
container with the same veid.

Signed-off-by: Glauber Costa <glommer@parallels.com>