vzctl won't work with ploop-1.4
But since we don't really want ploop-lib as a dependency
(because since vzctl is able to load ploop lib dynamically)
we just specify Conflicts with older version.

So, if ploop-lib < 1.5 is installed, vzctl installation will
complain (and yum will probably try to resolve it),
and if ploop-lib is not installed -- nothing will happen.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>