properly return errors in cgroup_init()
So far we were considering cgroup initialization to succeed as given.
This is of course not true, since it can fail if it is not present or
not mounted. Because we don't return an error in this case, vzctl will
happily continue, and fail in a worse way later on.

A bug arising from that is described at:

	http://bugzilla.openvz.org/show_bug.cgi?id=2372

and should be fixed by this patch.

Signed-off-by: Glauber Costa <glommer@parallels.com>
CC: Andrew Vagin <avagin@parallels.com>