print failures in ct_do_open directly to stderr
ct_do_open is called before any vzctl command, and has the purpose of
initializing all the data we need to proceed.

Some commands, like enter and status, will completely disable logging,
since they want a fixed output.

But errors in open are considered fatal. We would still like them to be
notified upwards. For this, we'll fprintf them to stderr, and make sure
to always return a meaningful vz error.

Signed-off-by: Glauber Costa <glommer@parallels.com>