src/Makefile.am: fix building with builddir != srcdir
From the original bug report:

> 1. git clone git://git.openvz.org/pub/vzctl
> 2. cd vzctl && ./autogen.sh
> 3. mkdir ../vzctl_build && cd ../vzctl_build
> 4. ../vzctl/configure && make
>
> make[2]: Entering directory `/home/max_posedon/OSS/vzctl_build/src'
>   CC       vzctl.o
> ../../vzctl/src/vzctl.c:26:21: fatal error: version.h: No such file or directory
> compilation terminated.
> make[2]: *** [vzctl.o] Error 1

This happens because version.h is generated by configure so it goes under
$(top_builddir) not $(top_srcdir). Therefore, the fix is add an appropriate
-I statement to CPPFLAGS.

I failed to fix it last time, together with similar issue fixed by
commit 5a428b1, because I was under a false impression that it is
the duty of autoconf to provide correct includes. In fact, it does
some trickery when AC_CONFIG_HEADERS is used but this is irrelevant.

http://bugzilla.openvz.org/2375

Reported-by: Maksim Melnikau <maxposedon@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>