configure: add ability to alter /vz path
With the help of a few previous patches, we can now
change /vz while doing ./configure.

Unfortunately, custom options available to ./configure
are limited to --with-something and --enable-something.
Neither works for this case, so we are using environment
variable:

	./configure vzdir=/some/dir

and (with the help of AC_ARG_VAR) vzdir will end up in Makefiles.

This fixes bug #421!

http://bugzilla.openvz.org/421

Signed-off-by: Kir Kolyshkin <kir@openvz.org>