Add OOM adjustments
Since kernel 042stab056 we have a kernel mechanism to auto-adjust process oom_score
based on it's name, parent name, and pid. This is very useful in multi-container
system, especially for old distros which are not oom-score aware.

OOM groups configuration is in /etc/vz/oom-groups.conf which is
applied to the kernel by /etc/init.d/vz script on start.

References: PSBM-12263, PSBM-13729
Signed-off-by: Kir Kolyshkin <kir@openvz.org>