properly propagate /var/lib/vzctl/veip dir
Otherwise we could run into a discrepancy then Makefile
thinks it's /usr/var and vzctl thinks it's /var.

I am not sure why it should be /usr/var and whoever on earth uses this,
but it's the same problem as with /etc vs /usr/etc (which is workarounded
in ./configure script).

While at it, also remove non-relevant comment about iptables
from scripts/vps-functions.in

Signed-off-by: Kir Kolyshkin <kir@openvz.org>