vzctl create: allow existing empty VE_PRIVATE
Previously, vzctl create checked if a directory for future CT's
private area exists -- and if it is, CT creation fails.

Apparently, it is problematic for some setups, when the directory
used for VE_PRIVATE must already be there.

Let's loose the check a but and allow for existing empty directory.

http://bugzilla.openvz.org/2450

Reported-by: Andrew Kornilov <akornilov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>