provide a upstream alternative for ub functions
The ub functions will not be ever supported in upstream kernel.
Still, they are called in a quite widespread fashion throughout the
code.

One possible course of action for this, is to skip the compilation
of ub.c, and at the same time provide an empty set of inline functions
in ub.h

[kir@openvz.org: remove #error, add comments #else/#endif */

Signed-off-by: Glauber Costa <glommer@parallels.com>