don't support vz-specific cpu functions if vz kernel is not present
We can't call any of the vz ioctls if we are not running in a
vz-compatible kernel. So if it is compile-time disabled, don't
compile them in.

We could do it in headers, but those functions will have upstream
similars, so we need them to be callable function objects, that will
just return a meaningful error in this case.

[kir@openvz.org: add comments to #else/#endif]

Signed-off-by: Glauber Costa <glommer@parallels.com>
(cherry picked from commit d335c6d2a34e03f8bdf8760aa67b537268f313ba)