configure: test for vz-kernel support in the target architecture
The include file "vzsyscalls.h" have a comprehensive list of vz-syscalls
and which architectures support them. If that file is included in a not
supported architecture, it will trigger an #error, which is a very
reliable way to determine that vz-specific code should not be compiled
in at all.

Since this places some new infrastructure in place for presence of
OpenVZ kernel support, this comes with the ability of disabling this
from the command line as well.

Signed-off-by: Glauber Costa <glommer@parallels.com>