vzlist: show if onboot field is unset
Field onboot was just yes/no (or true/false in json output), but
internally it can be in one of three states:

	0 NONE (i.e. unset)
	1 YES
	2 NO

This patchs adds vzlist support to display the 'unset' state,
as either '-' (in conventional output) or 'null' (in json output).

This functionality is needed by the following patch.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>