init.d/vz-gentoo: don't call tools by absolute path
Apparently in recent Gentoo the ip tool was moved from /sbin to /bin,
so this initscript stopped working.

Apparently runscript makes sure all the needed directories are in PATH
(I have tested with runscript from baselayout-1.12.11.1, pretty old one)
so we can just drop the /sbin/ prefix.

While at it, do it not only for ip, but also for modprobe and sysctl.

http://bugzilla.openvz.org/2477
https://bugs.gentoo.org/show_bug.cgi?id=444201

Reported-by: Andrei Vinogradov <spamslepnoga@inbox.ru>
Reported-by: Joakim <moonwalker@astro.nu>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>