vzevent: do umount/mount when rebooting CT
Apparently in some cases (like NFS mounts invoked by vzctl mount
scripts) vzevent should do umount/mount when restarting a container.
In general, this is the right thing to do anyway, and this is what vzctl
restart does. So, let's add umount to vzevent-reboot.

Also, since there is a hack to do reboot of Fedora 17 container in
vzevent-stop script, modify it accordingly to do umount before checking
if we need to reboot, not after.

https://bugzilla.openvz.org/2507

Reported-by: Pavel Snajdr <pavel.snajdr@relbit.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>