etc/dists/scripts/postcreate.sh: Fedora 18 support
For Fedora 18+, we need to set the following caps:

/usr/sbin/arping = cap_net_raw+ep
/usr/sbin/clockdiff = cap_net_raw+ep
/usr/bin/systemd-detect-virt = cap_dac_override,cap_sys_ptrace+ep

ALSO NOTE that in Fedora 17+ ping comes with CAP_NET_ADMIN which is
needed to use ping -m being non-root. But since CAP_NET_ADMIN is not
granted to a container by default, ping with this capability set
will not work at all. So we are NOT setting this cap for ping.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>