hooks_ct: fix pivot_root failure with mount namespaces.
Not only we don't need to call pivot_root when joining a mount namespace, we
can't do it. This is because the filesystem we currently are will become
invisible after we join the mount namespace. As a side effect, we are fully
protected because we will only have the filesystem view of the newly joined
namespace.

Because setns support for mount namespaces are not always present, we need
a flag do determine whether or not we've joined it. And because this is not
needed outside this scope at all, we can resort to a local only flag instead
of storing this information in the vps_handler

Signed-off-by: Glauber Costa <glommer@parallels.com>