vzmigrate: rewrite/fix IP address check
First, checking IPs one by one is not very efficient, so let's optimize,
rewriting this code to check all IPs at once, by making a A|B|C regexp
out of them.

Second, any IPs with netmasks present would not be detected -- for this
to work we need to remove netmasks from IPs if there are any.

This patch does both.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>