vzctl console: recognize ESC as a first character
After commit 88f582e, ESC is only recognized after Enter. While this is
a good thing in general, one possible scenario is a user runs vzctl
console, takes a look and tries to close it, typing ESC and dot.
This particular scenario is broken by the above commit.

I have checked how ssh works, and found that it does recognize ~. as a
first characters. So let's do the same, and fix the man page and the
variable name used accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>