ub: compile ub support for non vz kernels as well
Commit c9d9170b0 fixed a bug by not including the ub functions if VZ support
was not present in the running kernel, and replacing them with empty stubs.

This approach, however, proved to be too aggressive. We need to at least be
able to read and write the ub configuration from our conf files. The solution
proposed in this patch is to rework that so only the actual beancounter
operations are compiled out.

Signed-off-by: Glauber Costa <glommer@openvz.org>