pass parameters to open
Upstream containers running on kernels that support user namespaces would
benefit from being able to have early access to the container configuration.
This is because we would like user namespaces support to be either enabled or
disabled, and a lot of the actions we take can be potentially affected by this
configuration switch.

We do that by bundling it into vz_open, where we traditionally are setting all
the namespace related flags (based solely on the presence of it in the system).

The open handler in hooks_vz technically does not need to be modified. But I am
doing it out of consistency.

Signed-off-by: Glauber Costa <glommer@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>